+++ lib/screens/home_screen.dart
@@ -23,7 +23,7 @@ class _HomeScreenState extends State<HomeScreen> {
   final _ctrl = TextEditingController(text: _kSampleTexts.first);
   final _svc  = UDPipeService();
 
-  String _modelId = 'gsd';
+  String _modelId = 'german-gsd';
   UDPipeResult? _result;
   bool _processing = false;